Moved custom domains check to its own module and close session after every loop
